Re-integrate the `release/2` branch

+++ libcaf_net/caf/net/octet_stream/transport.cpp
@@ -413,13 +413,13 @@ protected:
   size_t max_consecutive_reads_ = defaults::middleman::max_consecutive_reads;
 
   /// Caches the write buffer size of the socket.
-  uint32_t max_write_buf_size_ = 0;
+  size_t max_write_buf_size_ = 0;
 
   /// Stores what the user has configured as read threshold.
-  uint32_t min_read_size_ = 0;
+  size_t min_read_size_ = 0;
 
   /// Stores what the user has configured as max. number of bytes to receive.
-  uint32_t max_read_size_ = 0;
+  size_t max_read_size_ = 0;
 
   /// Stores how many bytes are currently buffered, i.e., how many bytes from
   /// `read_buf_` are filled with actual data.
